Move useAddWallet to shared
